refactor(core): split ai/lobehub/gateway.ts into gateway/ modules
gateway.ts was 814 lines. Split into gateway/ by concern: json (shared
primitives), transport (HTTP error mapping, SSE parsing, trpc data
unwrapping), credentials (token/JWT helpers), models (model list mapping),
chatStream (payload building + SSE stream consumption), and client (the
WebApiLobeHubCloudGateway class). gateway.ts stays as a re-export barrel so
all existing import sites keep working unchanged. No logic changes.
