+++ packages/core/src/ai/lobehub/gateway/json.ts
@@ -1,0 +1,18 @@
+export type JsonObject = Record<string, unknown>;
+
+export function object(value: unknown): JsonObject | null {
+  return value && typeof value === 'object' && !Array.isArray(value) ? (value as JsonObject) : null;
+}
+
+export function number(value: unknown, fallback = 0): number {
+  if (typeof value === 'number' && Number.isFinite(value)) return value;
+  if (typeof value === 'string' && value.trim()) {
+    const parsed = Number(value);
+    if (Number.isFinite(parsed)) return parsed;
+  }
+  return fallback;
+}
+
+export function text(value: unknown): string | null {
+  return typeof value === 'string' && value ? value : null;
+}
